requirements.txt
pip~=25.3
PyJWT~=2.12.1
fastapi~=0.135.3
passlib~=1.7.4
pydantic-settings~=2.13.1
pydantic~=2.12.5